Added a class for placable items
