+++ src/main/java/mcvmcomputers/item/ItemList.java
@@ -3,10 +3,16 @@ package mcvmcomputers.item;
 import java.util.Arrays;
 import java.util.List;
 
+import mcvmcomputers.entities.EntityCRTScreen;
+import mcvmcomputers.entities.EntityFlatScreen;
+import mcvmcomputers.entities.EntityKeyboard;
+import mcvmcomputers.entities.EntityMouse;
+import mcvmcomputers.entities.EntityWallTV;
 import net.fabricmc.fabric.api.client.itemgroup.FabricItemGroupBuilder;
 import net.minecraft.block.Blocks;
 import net.minecraft.item.Item;
 import net.minecraft.item.Item.Settings;
+import net.minecraft.sound.SoundEvents;
 import net.minecraft.item.ItemGroup;
 import net.minecraft.item.ItemStack;
 import net.minecraft.util.Identifier;
@@ -20,12 +26,12 @@ public class ItemList {
 	public static final OrderableItem PC_CASE_SIDEPANEL = new ItemPCCaseSidepanel(new Settings().group(MOD_ITEM_GROUP_PARTS));
 	public static final OrderableItem ITEM_MOTHERBOARD = new OrderableItem(new Settings().group(MOD_ITEM_GROUP_PARTS), 4);
 	public static final OrderableItem ITEM_MOTHERBOARD64 = new OrderableItem(new Settings().group(MOD_ITEM_GROUP_PARTS), 8);
-	public static final OrderableItem ITEM_FLATSCREEN = new ItemFlatScreen(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS));
-	public static final OrderableItem ITEM_WALLTV = new ItemWallTV(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS));
-	public static final OrderableItem ITEM_CRTSCREEN = new ItemCRTScreen(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS));
+	public static final OrderableItem ITEM_FLATSCREEN = new PlacableOrderableItem(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS), EntityFlatScreen.class, SoundEvents.BLOCK_METAL_PLACE, 10);
+	public static final OrderableItem ITEM_WALLTV = new PlacableOrderableItem(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS), EntityWallTV.class, SoundEvents.BLOCK_METAL_PLACE, 14);
+	public static final OrderableItem ITEM_CRTSCREEN = new PlacableOrderableItem(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS), EntityCRTScreen.class, SoundEvents.BLOCK_METAL_PLACE, 10);
 	public static final OrderableItem ITEM_HARDDRIVE = new ItemHarddrive(new Settings().group(MOD_ITEM_GROUP_PARTS));
-	public static final OrderableItem ITEM_KEYBOARD = new ItemKeyboard(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS));
-	public static final OrderableItem ITEM_MOUSE = new ItemMouse(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS));
+	public static final OrderableItem ITEM_KEYBOARD = new PlacableOrderableItem(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS), EntityKeyboard.class, SoundEvents.BLOCK_METAL_PLACE, 4);
+	public static final OrderableItem ITEM_MOUSE = new PlacableOrderableItem(new Settings().group(MOD_ITEM_GROUP_PERIPHERALS), EntityMouse.class, SoundEvents.BLOCK_METAL_PLACE, 4);
 	public static final OrderableItem ITEM_RAM1G = new OrderableItem(new Settings().group(MOD_ITEM_GROUP_PARTS),2);
 	public static final OrderableItem ITEM_RAM2G = new OrderableItem(new Settings().group(MOD_ITEM_GROUP_PARTS),4);
 	public static final OrderableItem ITEM_RAM4G = new OrderableItem(new Settings().group(MOD_ITEM_GROUP_PARTS),8);
